+++ src/Cake.Core/IO/SpecialPath.cs
@@ -13,6 +13,6 @@
         /// <summary>
         /// The Windows folder.
         /// </summary>
-        Windows,
+        Windows,        
     }
 }